ZX Spectrum updates + GUI updates

+++ plugins/cpu/8080-cpu/src/test/java/net/emustudio/plugins/cpu/intel8080/suite/CpuRunnerImpl.java
@@ -76,7 +76,7 @@ public class CpuRunnerImpl extends CpuRunner<CpuImpl> {
 
     @Override
     public void setFlags(int mask) {
-        cpu.getEngine().flags |= mask;
+        cpu.getEngine().flags |= (short) mask;
     }
 
     @Override
